refactor(contracts): ModulePresentation 迁出 v1 契约文件 (#384)
`ModulePresentation` 与它依赖的 `ModuleStoryPage` 一直住在 v1 的
`contracts/module.py` 里，但它们描述的是「模组怎么呈现给玩家」——标题、简介、
人数、时长、导入分页——与 schema 版本无关，而且 `ModuleContentV3.presentation`
就是这个类型。下一步要删 v1 契约，这两个类必须先搬出来。

搬到 `contracts/module_v3.py`，`contracts/__init__.py` 的导出改从新位置取，
对外名字不变。v1 `ModuleContent` 自己也有 `presentation` 字段，所以
`module.py` 反过来 import 一次——`module_v3.py` 已经不再 import `module.py`，
不构成环，且这条反向依赖随下一步删 v1 契约一起消失。

零行为变化：`export_schema.py` 重跑后 `module-content-v3.schema.json` 无 diff，
两个套件计数不变（framework 335、backend 493 passed / 22 skipped）。
