+++ agent-collaboration-framework/collaboration_framework/contracts/__init__.py
@@ -125,8 +125,6 @@ from .module import (
     ModifyOperation,
     ModifyOperationSpec,
     ModuleContent,
-    ModulePresentation,
-    ModuleStoryPage,
     NarrativeDetailSpec,
     ObservableStateSpec,
     Operation,
@@ -218,6 +216,8 @@ from .module_v3 import (
     MatchOptionAuthorSpec,
     MatchQuestionSpec,
     ModuleContentV3,
+    ModulePresentation,
+    ModuleStoryPage,
     ModuleTimePolicySpec,
     NotCondition,
     PredicateCondition,
